chore: replace RecordList with RecordTable component and remove unused context for improved structure
